moved transform function into its own directory
